+++ HISTORY.rst
@@ -7,3 +7,13 @@ History
 ++++++++++++++++++
 
 * First release on PyPI.
+
+0.2.0 (2017-11-15)
+++++++++++++++++++
+* Updated AWS cloud model to reflect cloudbridge changes.
+* Minor bug fixes
+
+0.3.0 (2017-12-17)
+++++++++++++++++++
+* Moved azure resource group, storage account and vm_default_username to credentials
+* Changed status to beta